alert

+++ app/main.js
@@ -1,14 +1,21 @@
 import "./style.scss";
-import axios, { all } from "axios";
+import axios from "axios";
+import { z } from "zod";
+import moment from './node_modules/moment/dist/moment.js';
 
 let list = document.getElementById("list");
 
+const contentSchema = z.string().min(2).max(40).trim();
+
 function td(el) {
     return `<li class="li_list">
+  <div class="${el.done ? 'date_todo done' : 'date_todo'}">
+  <span>${moment(el.date).fromNow()}</span>
+  </div>
   <div class="content">
 
       <input type="text" data-id="${el.id}" value="${el.content}" class="${
-        el.done ? "done" : ""
+        el.done ? "done input_content_show" : "input_content_show"
     }" readonly>
       
   </div>
@@ -16,12 +23,9 @@ function td(el) {
       <button class="delete_todo" data-id="${el.id}">
           <i class="material-symbols-outlined"> delete </i>
       </button>
-      <button class="edit_todo" data-id="${el.id}" data-content="${el.content}">
-          <i class="material-symbols-outlined"> edit_document </i>
-      </button>
-      <button class="done_todo" data-id="${el.id}">
+      <button class="${el.done ? 'done_todo active' : 'done_todo'}" data-id="${el.id}">
           <i class="material-symbols-outlined"> ${
-              el.done ? "close" : "done"
+              el.done ? "bookmark" : "done"
           }  </i>
       </button>
   </div>
@@ -34,7 +38,7 @@ let todo_text = document.getElementById("todo_text");
 form.addEventListener("submit", (e) => {
     e.preventDefault();
 
-    if (todo_text.value.trim() !== "" || todo_text.value.trim() === null) {
+    if (contentSchema.safeParse(todo_text.value).success) {
         axios
             .post("http://localhost:3000/todo", { content: todo_text.value })
             .then((res) => {
@@ -51,7 +55,7 @@ form.addEventListener("submit", (e) => {
 });
 
 async function getData() {
-    const a = await axios("http://localhost:3000");
+    const a = await axios("http://localhost:3000/");
     return a;
 }
 
@@ -62,6 +66,8 @@ getData().then((res) => {
 function showData(arr) {
     list.innerHTML = "";
 
+    console.log(arr)
+
     arr.forEach((el) => {
         list.innerHTML += td(el);
     });
@@ -69,11 +75,25 @@ function showData(arr) {
 
 function deletTodo(id) {
     console.log(id);
-    axios.delete(`http://localhost:3000/todo/${id}`).then((res) => {
-        //console.log(res);
-        getData().then((r) => {
-            showData(r.data);
-        });
+    swal({
+        title: "Are you sure?",
+        text: "Lorem Ipsum is simply dummy text of the",
+        buttons: true,
+        dangerMode: true,
+    }).then((willDelete) => {
+        if (willDelete) {
+            axios.delete(`http://localhost:3000/todo/${id}`).then((res) => {
+                console.log(res.data.affectedRows);
+                if (res.data.affectedRows === 1) {
+                  getData().then((r) => {
+                    showData(r.data);
+                    //swal("Good job", "Lorem Ipsum is simply dummy text of the");
+                });
+                  
+                }
+                
+            });
+        }
     });
 }
 
@@ -87,10 +107,38 @@ function doneTodo(id) {
     });
 }
 
-function updateTodo(id, content) {
-    console.log(id);
-    document.querySelector(`[]`)
-    todo_text.value = content;
+function updateTodo(el, id) {
+    el.target.removeAttribute("readonly");
+    console.log(el.target);
+}
+
+function sendUpdateTodo(el, id) {
+    el.target.setAttribute("readonly", "");
+    //console.log(el.target.value);
+    const content = el.target.value;
+    if (contentSchema.safeParse(content).success) {
+        axios
+            .post(`http://localhost:3000/update/todo`, {
+                id: id,
+                content: content,
+            })
+            .then((res) => {
+                console.log(res);
+                getData().then((r) => {
+                    showData(r.data);
+                });
+            });
+    } else {
+        swal({
+            text: "Short text",
+            button: {
+                text: "Close",
+            },
+        });
+        getData().then((r) => {
+            showData(r.data);
+        });
+    }
 }
 
 const observer = new MutationObserver((obs) => {
@@ -104,15 +152,21 @@ const observer = new MutationObserver((obs) => {
         .forEach((el) =>
             el.addEventListener("click", (t) => doneTodo(t.target.dataset.id))
         );
+
     document
-        .querySelectorAll(".edit_todo")
+        .querySelectorAll(".input_content_show")
         .forEach((el) =>
             el.addEventListener("click", (t) =>
-                updateTodo(t.target.dataset.id, t.target.dataset.content)
+                updateTodo(t, t.target.dataset.id)
             )
         );
-
-    //console.log(obs.length - 1);
+    document
+        .querySelectorAll(".input_content_show")
+        .forEach((el) =>
+            el.addEventListener("blur", (t) =>
+                sendUpdateTodo(t, t.target.dataset.id)
+            )
+        );
 });
 
 observer.observe(list, { childList: true });
